refactor string extensions to refinements
