+++ examples/example.rb
@@ -1,3 +1,6 @@
+# This script demonstrates most functions of the Colorly DSL
+# Run it with `colorly tutorial.rb out.html`
+
 # Start a new color set
 title "One"
